working on type registration and inheritance
